micro: extract chat session title state
